Create mkvmerge-mediastripper.py
This is a makeshift python script to parse all the media file's mkvinfo JSON and identify tracks from the media that are not required, to save space.  
Expected input is a json file containing mkvmerge --identification-format json --identify outputs of each file in the media library in a file: mkvinfos.json  
Then the tracks are parsed and ONLY the Track IDs of the required audio/subtitle tracks are selected.  
Next it generates an mkvmerge command for each file, containing the required track IDs. This is saved as a script that can be run to rip only the required media tracks and save it in the destination.  
This has a lot of HARDCODED PATHS!!